Properly passthrough exit codes from grep
We want to keep grep's over all exit code style:

  0 = match found
  1 = no match found
  2 = error

xargs allows us to differ between exactly three types of error codes:

  0   = all processes exited with 0
  123 = at least one process exited with value between 1 and 125
  124 = the command exited with status 255

So let's map from grep to xargs:

  1 => 0
  0 => 1
  2 => 255

Then xargs maps:

  0   => 0
  1   => 123
  255 => 124

And then we map again from xargs to the final exit code:

  0   => 1
  123 => 0
  124 => 2

Yes, this is ugly. Especially the xargs output with -V respectively
--verbose. We might be able to unclutter this with capsulating that
"sh -c" call into a script on its own. But that's out of the scope of
this commit as this commit just fixes the bug as such and make s the
test suite pass again.